many changes
logger, fixes, deleteLevelUser and level comments

+++ src/features/comments/deleteAccComment.ts
@@ -1,6 +1,7 @@
 import { FastifyRequest, FastifyReply } from 'fastify'
 import { verifyGJP } from '../../lib/tools'
 import { query } from '../../lib/db'
+import Logger from '../../lib/logger'
 
 type Body = {
     accountID: number,
@@ -13,6 +14,8 @@ export default async function handler(req: FastifyRequest<{ Body: Body }>, rep: 
 
     if(!(await verifyGJP(req.body.accountID, req.body.gjp))) return -1
 
-    await query("DELETE FROM acc_comments WHERE accCommentID = ? LIMIT 1", [req.body.commentID])
+    await query("DELETE FROM acc_comments WHERE accCommentID = ? AND accountID = ? LIMIT 1", [req.body.commentID, req.body.accountID])
+    
+    Logger.event_delete('Account comment deleted')
     return 1
 }
